refactor(tier lists): tier list calculations go in the TierListSaga
PumbilityCensusSaga is deleted; its rebuild is now TierListSaga's fourth
consumer, beside the community, scores, and pass/PG builds — one saga owns the
tier-list calculations (owner, 2026-08-13). The command follows the family
naming (ProcessPumbilityTierListCommand), the Hangfire job becomes
process-pumbility-tier-list with a RemoveIfExists for the orphaned old id, and
the Admin button's method, test id, and snackbar say tier lists. PoolSize gets
one definition on PumbilityCohortKeys, shared by the writer's membership gate
and the reader's resolution.

The saga tests move to TierListSagaPumbilityTests (the TierListSagaStaticsTests
split precedent) and construct the real TierListSaga.
