--- ScoreTracker/ScoreTracker.ChartIntelligence/Contracts/Messages/ProcessPumbilityCensusCommand.cs
@@ -1,9 +1,0 @@
-using ScoreTracker.SharedKernel.Enums;
-
-namespace ScoreTracker.ChartIntelligence.Contracts.Messages;
-
-// Single-mix per message (plan doc): replaying one mix's recompute never touches the other's.
-[ExcludeFromCodeCoverage]
-public sealed record ProcessPumbilityCensusCommand(MixEnum Mix = MixEnum.Phoenix)
-{
-}
