+++ ScoreTracker/ScoreTracker/Pages/Admin/Admin.razor
@@ -49,13 +49,13 @@
                data-testid="@($"reprice-ratings-{mix}")">@L["Re-price {0} ratings", mix.GetName()]</MudButton>
 }
 
-@* The PUMBILITY lens reads a nightly census, so a fresh deploy shows nothing until the job
+@* The PUMBILITY lens reads a nightly count, so a fresh deploy shows nothing until the job
    first runs — and locally PreventRecurringJobs parks it on a yearly cron, where the Hangfire
    dashboard cannot trigger it either. Safe to re-run: the job rewrites each folder wholesale. *@
 @foreach (var mix in RepriceableMixes)
 {
-    <MudButton Color="Color.Warning" Variant="Variant.Filled" OnClick="@(() => RebuildPumbilityCensus(mix))"
-               data-testid="@($"rebuild-pumbility-census-{mix}")">@L["Rebuild {0} PUMBILITY tier lists", mix.GetName()]</MudButton>
+    <MudButton Color="Color.Warning" Variant="Variant.Filled" OnClick="@(() => RebuildPumbilityTierLists(mix))"
+               data-testid="@($"rebuild-pumbility-tier-lists-{mix}")">@L["Rebuild {0} PUMBILITY tier lists", mix.GetName()]</MudButton>
 }
 
 <MudCard>
@@ -252,10 +252,10 @@
             Severity.Success);
     }
 
-    private async Task RebuildPumbilityCensus(MixEnum mix)
+    private async Task RebuildPumbilityTierLists(MixEnum mix)
     {
-        await Bus.Publish(new ProcessPumbilityCensusCommand(mix));
-        Snackbar.Add(L["Rebuilding the {0} PUMBILITY census. It takes a few minutes.", mix.GetName()].Value,
+        await Bus.Publish(new ProcessPumbilityTierListCommand(mix));
+        Snackbar.Add(L["Rebuilding the {0} PUMBILITY tier lists. It takes a few minutes.", mix.GetName()].Value,
             Severity.Success);
     }
